+++ e2e/commands/import.e2e.js
@@ -122,14 +122,6 @@ describe('bit import', function () {
       });
       describe('when imported', () => {
         let localConsumerFiles;
-        before(() => {
-          helper.reInitLocalScope();
-          helper.addRemoteScope();
-          const output = helper.importComponent('imprel/impreldist');
-          expect(output.includes('successfully imported one component')).to.be.true;
-          expect(output.includes('imprel/imprel')).to.be.true;
-          localConsumerFiles = helper.getConsumerFiles();
-        });
         const imprelDir = path.join('components', 'imprel', 'impreldist');
         const expectedLocationImprel = path.join(imprelDir, 'imprel.js');
         const expectedLocationImprelSpec = path.join(imprelDir, 'imprel.spec.js');
@@ -137,19 +129,11 @@ describe('bit import', function () {
         const expectedLocationImprelDist = path.join(imprelDir, 'dist', 'imprel.js');
         const expectedLocationImprelSpecDist = path.join(imprelDir, 'dist', 'imprel.spec.js');
         const expectedLocationMyUtilDist = path.join(imprelDir, 'dist', 'utils', 'myUtil.js');
-        it('should write the internal files according to their relative paths', () => {
-          expect(localConsumerFiles).to.include(expectedLocationImprel);
-          expect(localConsumerFiles).to.include(expectedLocationImprelSpec);
-          expect(localConsumerFiles).to.include(expectedLocationMyUtil);
-        });
-        it('should write the dist files in the component root dir', () => {
-          expect(localConsumerFiles).to.include(expectedLocationImprelDist);
-          expect(localConsumerFiles).to.include(expectedLocationImprelSpecDist);
-          expect(localConsumerFiles).to.include(expectedLocationMyUtilDist);
-        });
-        describe('when a project is cloned somewhere else as IMPORTED', () => {
+        describe('with --dist flag', () => {
           before(() => {
-            helper.mimicGitCloneLocalProject();
+            helper.reInitLocalScope();
+            helper.addRemoteScope();
+            helper.importComponent('imprel/impreldist --dist');
             localConsumerFiles = helper.getConsumerFiles();
           });
           it('should write the internal files according to their relative paths', () => {
@@ -161,6 +145,73 @@ describe('bit import', function () {
             expect(localConsumerFiles).to.include(expectedLocationImprelDist);
             expect(localConsumerFiles).to.include(expectedLocationImprelSpecDist);
             expect(localConsumerFiles).to.include(expectedLocationMyUtilDist);
+          });
+          describe('when a project is cloned somewhere else as IMPORTED', () => {
+            before(() => {
+              helper.mimicGitCloneLocalProject(true);
+              localConsumerFiles = helper.getConsumerFiles();
+            });
+            it('should write the internal files according to their relative paths', () => {
+              expect(localConsumerFiles).to.include(expectedLocationImprel);
+              expect(localConsumerFiles).to.include(expectedLocationImprelSpec);
+              expect(localConsumerFiles).to.include(expectedLocationMyUtil);
+            });
+            it('should write the dist files in the component root dir', () => {
+              expect(localConsumerFiles).to.include(expectedLocationImprelDist);
+              expect(localConsumerFiles).to.include(expectedLocationImprelSpecDist);
+              expect(localConsumerFiles).to.include(expectedLocationMyUtilDist);
+            });
+          });
+        });
+        describe('with --dist flag when dist is set to a non-default directory', () => {
+          before(() => {
+            helper.reInitLocalScope();
+            helper.addRemoteScope();
+            helper.modifyFieldInBitJson('dist', { target: 'another-dist' });
+            helper.importComponent('imprel/impreldist --dist');
+            localConsumerFiles = helper.getConsumerFiles();
+          });
+          it('should write the dist files according to the new dist-target set in bit.json', () => {
+            const newLocationImprelDist = path.join(imprelDir, 'another-dist', 'imprel.js');
+            const newLocationImprelSpecDist = path.join(imprelDir, 'another-dist', 'imprel.spec.js');
+            const newLocationMyUtilDist = path.join(imprelDir, 'another-dist', 'utils', 'myUtil.js');
+            expect(localConsumerFiles).to.include(newLocationImprelDist);
+            expect(localConsumerFiles).to.include(newLocationImprelSpecDist);
+            expect(localConsumerFiles).to.include(newLocationMyUtilDist);
+          });
+        });
+        describe('without --dist flag', () => {
+          before(() => {
+            helper.reInitLocalScope();
+            helper.addRemoteScope();
+            helper.importComponent('imprel/impreldist');
+            localConsumerFiles = helper.getConsumerFiles();
+          });
+          it('should write the internal files according to their relative paths', () => {
+            expect(localConsumerFiles).to.include(expectedLocationImprel);
+            expect(localConsumerFiles).to.include(expectedLocationImprelSpec);
+            expect(localConsumerFiles).to.include(expectedLocationMyUtil);
+          });
+          it('should not write the dist files in the component root dir', () => {
+            expect(localConsumerFiles).to.not.include(expectedLocationImprelDist);
+            expect(localConsumerFiles).to.not.include(expectedLocationImprelSpecDist);
+            expect(localConsumerFiles).to.not.include(expectedLocationMyUtilDist);
+          });
+          describe('when a project is cloned somewhere else as IMPORTED', () => {
+            before(() => {
+              helper.mimicGitCloneLocalProject();
+              localConsumerFiles = helper.getConsumerFiles();
+            });
+            it('should write the internal files according to their relative paths', () => {
+              expect(localConsumerFiles).to.include(expectedLocationImprel);
+              expect(localConsumerFiles).to.include(expectedLocationImprelSpec);
+              expect(localConsumerFiles).to.include(expectedLocationMyUtil);
+            });
+            it('should not write the dist files in the component root dir', () => {
+              expect(localConsumerFiles).to.not.include(expectedLocationImprelDist);
+              expect(localConsumerFiles).to.not.include(expectedLocationImprelSpecDist);
+              expect(localConsumerFiles).to.not.include(expectedLocationMyUtilDist);
+            });
           });
         });
       });
@@ -932,7 +983,7 @@ describe('bit import', function () {
       helper.exportAllComponents();
       helper.reInitLocalScope();
       helper.addRemoteScope();
-      helper.importComponent('bar/foo');
+      helper.importComponent('bar/foo --dist');
       localConsumerFiles = helper.getConsumerFiles();
     });
     it('should keep the original directory structure of the main component', () => {