+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- prevent overriding index file if exists 
+- write dist files for imported components according to dist.target configuration in bit.json
+- do not write dist files for imported components, unless `--dist` flag is used
+- prevent overriding index file if exists
 
 ## [0.11.2-dev.1] - 2017-12-04